modularizing index page into components
