+++ app/components/Headline/Headline.module.css
@@ -1,0 +1,32 @@
+.content {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 2rem;
+}
+
+.content__headline {
+  text-align: center;
+  margin-bottom: 2rem;
+}
+
+.content__subheadline {
+  font-size: var(--h3-font-size);
+  text-align: center;
+}
+
+@media (min-width: 1024px) {
+  .content__headline {
+    max-width: 900px;
+  }
+
+  .content__subheadline {
+    max-width: 592px;
+  }
+
+  .content__headline {
+    font-size: var(--xtra-big-font-size);
+    line-height: 100px;
+  }
+}
